chore(review-gaps): dead-column + boolean-mode + validation + modal/DX fixes
Closes the non-rename findings from the role-based reviews
(docs/superpowers/reviews):

- Drop dead `tenants.nachi_number` (unwired since introduction; zero
  reads/writes) and its vestigial `admin/settings.ts` validation field.
- `qbo_sync_errors.resolved` -> `integer { mode: 'boolean' }` (+ the four
  read/write sites); naming-audit C5.
- `UpdateInspectionSchema.price` now `.int()` to match the create schema and
  the cents convention; naming-audit C4 ride-along.
- Regenerate the single `0000_baseline.sql` from schema (db:check clean, 71=71)
  and drop the stale `nachi_number` column from the three workers inline
  `tenants` DDLs so they stay in sync with the schema.
- New-inspection modal: cap at `max-h-[90vh]` with a scrollable body so a tall
  step never pushes the footer off-screen on short viewports; field UX review.
- `prepare` now also generates `server/version.ts` (guarded) and a `bootstrap`
  script (gen-version + vendor:copy) prevents fresh-checkout type-check false-reds.

+++ server/lib/validations/inspection/crud.ts
@@ -103,7 +103,7 @@ export const UpdateInspectionSchema = z.object({
     clientEmail: z.string().email().optional().nullable().openapi({ example: 'john@example.com' }).describe('TODO describe clientEmail field for the OpenInspection MCP integration'),
     date: z.string().datetime().optional().openapi({ example: '2024-03-20T10:00:00Z' }).describe('TODO describe date field for the OpenInspection MCP integration'),
     inspectorId: z.string().uuid().optional().openapi({ example: '550e8400-e29b-41d4-a716-446655440001' }).describe('TODO describe inspectorId field for the OpenInspection MCP integration'),
-    price: z.number().min(0).optional().openapi({ example: 450 }).describe('TODO describe price field for the OpenInspection MCP integration'),
+    price: z.number().int().min(0).optional().openapi({ example: 450 }).describe('TODO describe price field for the OpenInspection MCP integration'),
     status: z.enum(INSPECTION_STATUSES).optional().openapi({ example: 'completed' }).describe('TODO describe status field for the OpenInspection MCP integration'),
     paymentRequired:   z.boolean().optional().openapi({ example: false }).describe('TODO describe paymentRequired field for the OpenInspection MCP integration'),
     agreementRequired: z.boolean().optional().openapi({ example: false }).describe('TODO describe agreementRequired field for the OpenInspection MCP integration'),
